Search for all @see tags on all methods of same name for related.
If the @webref tag is written above a different version of a method
from the version that has @see tags, javadoc reports that method as
(correctly) having no @see tags. We now search through all the versions
of the method and aggregate all the @see tags for the Processing
related section.
Fixes gh 293